[Minor] improvement(docs): Remove support for data type `map` and `interval_year` for Lance table in docs. (#9494)
### What changes were proposed in this pull request?

This pull request updates the documentation for Lance's support of
Apache Arrow types, clarifying which types are not supported. The most
important changes are clarifications in the type mapping table to
indicate unsupported types.

Documentation updates:

* Updated the type mapping table in
`docs/lakehouse-generic-lance-table.md` to clarify that the `Map` type
is not supported by Lance, instead of mapping it to Arrow's `Map` type.
* Updated the type mapping table to indicate that `Interval_year` is not
supported by Lance, instead of mapping it to Arrow's
`Interval(YearMonth)` type.

### Why are the changes needed?

Data type `map` and `interval(yearmonth)` are not supported by Lance in
release 0.39.0


### Does this PR introduce _any_ user-facing change?

N/A

### How was this patch tested?

Test locally.